% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data for HVP Model}
\format{
A list with the following elements:
\describe{
\item{tau}{Causal effect of the exposure on the outcome.}
\item{varb}{Genetic variance of the exposure variable.}
\item{covg_hat}{Genetic covariance between exposure and outcome.}
\item{varg}{Genetic variance of the outcome variable.}
\item{v_tau}{Variance of the causal effect (tau).}
\item{v_varb}{Variance of the exposure variable.}
\item{v_varg}{Variance of the outcome variable.}
\item{v_covg_hat}{Variance of the genetic covariance (covg_hat).}
\item{cov_varg_covg_hat}{Covariance between varg and covg_hat.}
\item{cov_varb_covg_hat}{Covariance between varb and covg_hat.}
}
}
\usage{
data(example_data)
}
\description{
A dataset to demonstrate the usage of the genetic correlation estimation function in the HVP model.
}
\examples{
data(example_data)
print(example_data)
}
\keyword{datasets}
