% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVP.R
\docType{class}
\name{hvp-class}
\alias{hvp-class}
\title{HVP results class}
\description{
An S4 class to store the results from Hierarchical variance partitioning (HVP).
}
\section{Slots}{

\describe{
\item{\code{HVP}}{numeric indicating the proportion of variance associated with
batch effects.}

\item{\code{sum.squares}}{matrix containing sum of squares between batches and
total sum of squares for all features.}

\item{\code{p.value}}{optional numeric of P-value from permutation test.}

\item{\code{null.distribution}}{optional numeric vector of null distribution of
HVP values.}
}}

