% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVP.R
\name{hvp}
\alias{hvp}
\title{Genetic Correlation Estimation}
\usage{
hvp(
  varg,
  varb,
  covg_hat,
  v_varg,
  v_varb,
  v_covg_hat,
  cov_varg_varb,
  cov_varg_covg_hat,
  cov_varb_covg_hat,
  tau
)
}
\arguments{
\item{varg}{Genetic variance of the outcome variable estimated in a univariate context.}

\item{varb}{Genetic variance of the exposure variable estimated in a univariate context.}

\item{covg_hat}{Genetic covariance between the exposure variable ('varb') and the outcome variable ('varg'), as estimated by the conventional bivariate GREML method using genome-wide SNPs.}

\item{v_varg}{Variance of 'varg' from the information matrix generated by the conventional bivariate GREML method, or the square of the standard error of 'varg'.}

\item{v_varb}{Variance of 'varb' from the information matrix generated by the conventional bivariate GREML method, or the square of the standard error of 'varb'.}

\item{v_covg_hat}{Variance of 'covg_hat' from the information matrix generated by the conventional bivariate GREML method, or the square of the standard error of 'covg_hat'.}

\item{cov_varg_varb}{Covariance between 'varb' and 'varg', derived from the information matrix.}

\item{cov_varg_covg_hat}{Covariance between 'varg' and the genetic covariance ('covg_hat') between 'varb' and 'varg', derived from the information matrix.}

\item{cov_varb_covg_hat}{Covariance between 'varb' and the genetic covariance ('covg_hat') between 'varb' and 'varg', derived from the information matrix.}

\item{tau}{The causal effect of the exposure on the outcome, as estimated by a robust Mendelian Randomization (MR) method.}
}
\value{
A matrix containing:
\item{Estimate}{Corrected estimates of 'Heritability', 'Genetic Covariance', and 'Genetic Correlation'.}
\item{Standard Error}{Standard errors of the corrected estimates.}
\item{P_value}{P-values from chi-square tests for the corrected estimates.}
}
\description{
The Horizontal and Vertical Pleiotropy (HVP) model applied to individual-level data. This method disentangles horizontal from vertical pleiotropy effects, providing unbiased genetic correlation estimates specifically attributed to horizontal pleiotropy. It corrects the genetic correlation derived from bivariate REML.
}
