% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVP.R
\name{plot,hvp,missing-method}
\alias{plot,hvp,missing-method}
\title{Plot results of permutation test}
\usage{
\S4method{plot}{hvp,missing}(x, y, ...)
}
\arguments{
\item{x}{hvp S4 class containing HVP results after permutation testing.}

\item{y}{ignored argument for compatibility with generic plot function.}

\item{...}{ignored argument for compatibility with generic plot function.}
}
\value{
ggplot object of null distribution of permutation test.
}
\description{
Plot results of permutation test
}
\details{
Plots the null distribution of the permutation test.
}
