% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RVP.R
\name{RVP}
\alias{RVP}
\alias{RVP.default}
\alias{RVP.Seurat}
\alias{RVP.SummarizedExperiment}
\title{Recursive variance partitioning (RVP)}
\usage{
RVP(x, ...)

\method{RVP}{default}(X, batch, cls = NULL, nperm = 0, use.sparse = FALSE)

\method{RVP}{Seurat}(seu, batchname, classname = NULL, slot = "data", ...)

\method{RVP}{SummarizedExperiment}(se, batchname, classname = NULL, assayname = NULL, ...)
}
\arguments{
\item{x}{object to calculate RVP for.}

\item{...}{optional arguments to pass to `RVP.default`}

\item{X}{data frame or matrix with dim (nfeatures, nsamples).}

\item{batch}{vector, indicating the batch information of samples.}

\item{cls}{vector or list of vectors with class information of samples.}

\item{nperm}{numeric indicating number of permutations to simulate in the
Monte Carlo permutation test. We recommend a value no less than 1000.
By default, no permutation test is performed.}

\item{use.sparse}{logical indicating whether to use sparse matrices when
 computing RVP. N.B. Using sparse matrices may lead to slight increase
in run time.}

\item{seu}{Seurat object.}

\item{batchname}{character, name of column in metadata indicating batch.}

\item{classname}{character, name of column/s in metadata indicating class.}

\item{slot}{character, name of slot in assay to use. E.g. counts.}

\item{se}{`SummarizedExperiment`/`SingleCellExperiment` object.
`SingleExperiment` class inherits from the `SummarizedExperiment` class.}

\item{assayname}{character, name of assay to use. By default the first
assay is used.}
}
\value{
List containing the following components:
  \describe{
    \item{`RVP`}{the proportion of variance associated with batch effects.}
    \item{`sum.squares`}{matrix of sum of squares between batch and total
      sum of squares for all features.}
    \item{`p.value`}{p-value of permutation test}
    \item{`null.distribution`}{numeric, null distribution of RVP values.}
  }
}
\description{
`RVP` calculates the proportion of variance associated with batch effects
in a data set (the "RVP" value of a data set). To determine whether batch
effects are statistically significant in a data set, a permutation test can
be performed by setting `nperm` to a number above 100. `RVP` is
an S3 generic function; methods can be added for new classes. S3 methods
for class: array-like objects (default), `SummarizedExperiment`, 
`SingleCellExperiment` and `Seurat` are provided.
}
\author{
Wei Xin Chan
}
