% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitCols}
\alias{splitCols}
\title{Splits subsettable objects according to their columns}
\usage{
splitCols(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{subsettable object to be split}

\item{f}{vector or list of vectors indicating the grouping of columns}

\item{drop}{logical indicating if levels that do not occur should be dropped}

\item{...}{optional arguments to [split()]}
}
\value{
List of objects split by columns
}
\description{
Splits subsettable objects according to their columns
}
\examples{

X <- matrix(1:60, 10, 6)
cond <- rep(1:3, each = 2)
splitCols(X, cond)

}
