% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_microarray}
\alias{simulate_microarray}
\title{Simulate log-transformed gene expression microarray data}
\usage{
simulate_microarray(
  m,
  crosstab,
  delta = 1,
  gamma = 0.5,
  phi = 0.2,
  c = 10,
  d = 6,
  epsilon = 0.5,
  kappa = 0.2,
  a = 40,
  b = 5,
  dropout = FALSE,
  r = 2,
  s = -6,
  seed = NA
)
}
\arguments{
\item{m}{number of features.}

\item{crosstab}{matrix of contingency table specifying number of samples in
each class-batch condition, with classes as rows and batches as columns.}

\item{delta}{magnitude of additive batch effects.}

\item{gamma}{magnitude of multiplicative batch effects. Variance parameter
of Normal distribution modelling batch effects across samples in a batch.}

\item{phi}{percentage of differentially expressed features.}

\item{c}{shape parameter of Gamma distribution modelling log fold-change.}

\item{d}{rate parameter of Gamma distribution modelling log fold-change.}

\item{epsilon}{magnitude of feature-wise variation.}

\item{kappa}{magnitude of sample-specific variation (scaling factor).}

\item{a}{shape parameter of Gamma distribution modelling basal expression.}

\item{b}{rate parameter of Gamma distribution modelling basal expression.}

\item{dropout}{logical indicating whether to perform dropout}

\item{r}{inverse scale parameter of the sigmoid function.}

\item{s}{midpoint parameter of the sigmoid function.}

\item{seed}{numeric specifying random seed. Defaults to no seed.}
}
\value{
A list containing the following components:
  \describe{
    \item{`X`}{matrix with dimensions `(m, n)` of log expression values.}
    \item{`metadata`}{data frame with `n` rows of sample metadata.}
    \item{`diff.features`}{character vector, names of differentially
      expressed features.}
    \item{`W`}{matrix with dimensions `(m, n)` of log expression values
      without addition of batch effect term.}
    \item{`batch.terms`}{matrix with dimensions (m, n) of batch effect terms.}
    \item{`class.logfc`}{matrix of mean log fold change for each feature
      for classes.}
    \item{`batch.logfc`}{matrix of mean log fold change for each features
      for batches.}
    \item{`params`}{list of parameters supplied.}
  }
}
\description{
Simulate log-transformed gene expression microarray data
}
\author{
Wei Xin Chan
}
